+++ iotdb-core/datanode/src/main/java/org/apache/iotdb/db/pipe/receiver/thrift/IoTDBThriftReceiverAgent.java
@@ -19,7 +19,6 @@
 
 package org.apache.iotdb.db.pipe.receiver.thrift;
 
-import org.apache.iotdb.db.conf.IoTDBDescriptor;
 import org.apache.iotdb.db.pipe.connector.protocol.IoTDBConnectorRequestVersion;
 import org.apache.iotdb.db.queryengine.plan.analyze.IPartitionFetcher;
 import org.apache.iotdb.db.queryengine.plan.analyze.schema.ISchemaFetcher;
@@ -28,12 +27,8 @@ import org.apache.iotdb.rpc.TSStatusCode;
 import org.apache.iotdb.service.rpc.thrift.TPipeTransferReq;
 import org.apache.iotdb.service.rpc.thrift.TPipeTransferResp;
 
-import org.apache.commons.io.FileUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.io.File;
-import java.io.IOException;
 
 public class IoTDBThriftReceiverAgent {
 
@@ -89,25 +84,6 @@ public class IoTDBThriftReceiverAgent {
     if (receiver != null) {
       receiver.handleExit();
       receiverThreadLocal.remove();
-    }
-  }
-
-  public void cleanPipeReceiverDir() {
-    final File receiverFileDir =
-        new File(IoTDBDescriptor.getInstance().getConfig().getPipeReceiverFileDir());
-
-    try {
-      FileUtils.deleteDirectory(receiverFileDir);
-      LOGGER.info("Clean pipe receiver dir {} successfully.", receiverFileDir);
-    } catch (Exception e) {
-      LOGGER.warn("Clean pipe receiver dir {} failed.", receiverFileDir, e);
-    }
-
-    try {
-      FileUtils.forceMkdir(receiverFileDir);
-      LOGGER.info("Create pipe receiver dir {} successfully.", receiverFileDir);
-    } catch (IOException e) {
-      LOGGER.warn("Create pipe receiver dir {} failed.", receiverFileDir, e);
     }
   }
 }
