+++ iotdb-core/datanode/src/main/java/org/apache/iotdb/db/service/DataNode.java
@@ -62,7 +62,6 @@ import org.apache.iotdb.db.conf.rest.IoTDBRestServiceDescriptor;
 import org.apache.iotdb.db.consensus.DataRegionConsensusImpl;
 import org.apache.iotdb.db.consensus.SchemaRegionConsensusImpl;
 import org.apache.iotdb.db.pipe.agent.PipeAgent;
-import org.apache.iotdb.db.pipe.receiver.airgap.IoTDBAirGapReceiverAgent;
 import org.apache.iotdb.db.protocol.client.ConfigNodeClient;
 import org.apache.iotdb.db.protocol.client.ConfigNodeClientManager;
 import org.apache.iotdb.db.protocol.client.ConfigNodeInfo;
@@ -99,7 +98,6 @@ import java.nio.ByteBuffer;
 import java.nio.charset.Charset;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 
@@ -188,9 +186,6 @@ public class DataNode implements DataNodeMBean {
 
       // Setup rpc service
       setUpRPCService();
-
-      // Setup pipe air gap service
-      setUpPipeAirGapService();
 
       // Serialize mutable system properties
       IoTDBStartCheck.getInstance().serializeMutableSystemPropertiesIfNecessary();
@@ -558,13 +553,6 @@ public class DataNode implements DataNodeMBean {
     registerManager.register(PipeAgent.runtime());
   }
 
-  /** Set up air gap service if pipe air gap receiver is enabled. */
-  private void setUpPipeAirGapService() {
-    if (PipeConfig.getInstance().getPipeAirGapReceiverEnabled()) {
-      Optional.ofNullable(PipeAgent.receiver().airGap()).ifPresent(IoTDBAirGapReceiverAgent::start);
-    }
-  }
-
   /** Set up RPC and protocols after DataNode is available */
   private void setUpRPCService() throws StartupException {
     // Start InternalRPCService to indicate that the current DataNode can accept cluster scheduling
@@ -902,6 +890,9 @@ public class DataNode implements DataNodeMBean {
     }
     if (IoTDBRestServiceDescriptor.getInstance().getConfig().isEnableRestService()) {
       registerManager.register(RestService.getInstance());
+    }
+    if (PipeConfig.getInstance().getPipeAirGapReceiverEnabled()) {
+      registerManager.register(PipeAgent.receiver().airGap());
     }
   }
